requirements.txt
PyQt5
pyserial
glob3
os-sys